Allow to also play specified playlists and tracks
Every item on the command line is a so called track group. Reordering does not affect tracks across track group boundaries.
The arguments genre and time are useless if urls are specified. Use the special form "trending:genre:time" instead.

This also includes a lot of refactoring and smaller improvements.

+++ src/ui/log.rs
@@ -1,7 +1,7 @@
 use terminal_size::{terminal_size, Width};
 
+use super::HalkaraUi;
 use crate::audius::TrackGroup;
-use crate::ui::HalkaraUi;
 
 pub struct Log;
 
@@ -38,7 +38,7 @@ impl HalkaraUi for Log {
     fn cleanup(&self) {}
 }
 
-fn print_rank(rank: u8) {
+fn print_rank(rank: usize) {
     let term_size = terminal_size();
     let width = if let Some((Width(w), _)) = term_size {
         w
